tic_tac_toe/requirement.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==3.0.0
channels-redis==3.2.0
constantly==15.1.0
cryptography==41.0.3
daphne==3.0.2
Django==4.2.5
djangorestframework==3.14.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
numpy==1.26.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.5.2
pyOpenSSL==23.2.0
pytz==2023.3.post1
redis==5.0.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
zope.interface==6.0
